+++ shared/constants/types/dev.js
@@ -1,0 +1,13 @@
+// @flow
+
+export type DebugConfig = {|
+  dumbFilter: string,
+  dumbFullscreen: boolean,
+  dumbIndex: number,
+|}
+
+export type State = {
+  debugConfig: DebugConfig,
+  debugCount: number,
+  hmrReloading: boolean,
+}
