+++ shared/reducers/index.js
@@ -1,5 +1,4 @@
 // @flow
-// import * as Constants from '../constants/dev'
 import chat from './chat'
 import config from './config'
 import dev from './dev'
